ポーズ処理追加
