+++ Assets/Prefabs/UI/MiniGame.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 76ce66624528dda4591dab8ad787316f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
