+++ Assets/Prefabs/UI/MiniGame/MiniGamePauseUI.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a023d2d33bd756349adc84134e585e59
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
