+++ Assets/Scenes/CircuitGame/MiniGamePauseUI.controller
@@ -58,7 +58,7 @@ AnimatorController:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: CircuitPauseUI
+  m_Name: MiniGamePauseUI
   serializedVersion: 5
   m_AnimatorParameters: []
   m_AnimatorLayers:
