+++ Assets/Scenes/CircuitGame/MiniGamePauseUIFadeOut.anim
@@ -6,7 +6,7 @@ AnimationClip:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: CircuitPauseUIFadeOut
+  m_Name: MiniGamePauseUIFadeOut
   serializedVersion: 6
   m_Legacy: 0
   m_Compressed: 0
